Separated the themes dropdown menu logic into its own script and cleaned up the settings dropdown scripts
